[#11]: Restructure modules in more natural way
Problem:
There was a question where to put `profile` endpoints, and more natural
way is not to put it in separate group, but under `accounts`.
So, to do it in nice way - we need to make some order.

Solution:
Split types and schema for `signIn` and general `account`, restructure
with dirs - so `Account` have under it `Profile` and `SignIn` now.
Make related changes to tests.
