+++ backend/src/Union/Account/Server.hs
@@ -11,101 +11,26 @@ module Union.Account.Server
 
 import Relude
 
-import Network.Socket (SockAddr)
-import Servant ((:>), JSON, Post, ReqBody)
-import Servant.API
-  ( Description
-  , Header
-  , NamedRoutes
-  , NoContent(..)
-  , PostCreated
-  , RemoteHost
-  , Summary
-  )
+import Servant ((:>))
+import Servant.API (NamedRoutes)
 import Servant.API.Generic ((:-))
 
-import Core.Error (throwError, throwOnNothingM)
-import Core.Logger (Severity(..), logDebug, logInfo)
-import Core.Sender (isPhoneValid)
-
-import Union.Account.Schema (ConfirmationScope(..), UserAgent)
-import Union.Account.Service
-  (createAccount, findAccount, sendConfirmationCode, signIn)
-import Union.Account.Types (RequestCodeReq(..), SignInReq(..), SignInResp(..))
-import Union.App.Env (Union, WithError, WithJwt, WithLog, WithSender)
-import Union.App.Error (Error(..))
+import Union.Account.Profile.Server (ProfileAPI, profileEndpoints)
+import Union.Account.SignIn.Server (SignInAPI, signInEndpoints)
+import Union.App.Env (Union)
 
 
 -- | Helper type to represent Account API in terms of Servant.
-type AccountAPI = NamedRoutes AccountEndpoints
+type AccountAPI = "accounts" :> NamedRoutes AccountEndpoints
 
 -- | Represents API related to account.
 data AccountEndpoints mode = AccountEndpoints
-  { _requestCode :: mode
-      :- "accounts"
-      :> "signIn"
-      :> "request"
-      :> Summary "Request code to Sign In"
-      :> Description
-        "First step in Sign In process. At this step user requests confirmation \
-        \code, which we will send to provided phone. If there is no account with\
-        \ such phone - we will create it and send code then."
-      :> ReqBody '[JSON] RequestCodeReq
-      :> PostCreated '[JSON] NoContent
-  , _signIn :: mode
-      :- "accounts"
-      :> "signIn"
-      :> Summary "Sign In to account"
-      :> Description
-        "Second step in Sign In process. At this step user confirms that account\
-        \ belongs to him - we expect phone number and confirmation code, which \
-        \was sent at previous step. As a result - user will get JWT token."
-      :> RemoteHost
-      :> Header "User-Agent" UserAgent
-      :> ReqBody '[JSON] SignInReq
-      :> Post '[JSON] SignInResp
-  } deriving stock (Generic)
+  { _signIn  :: mode :- SignInAPI
+  , _profile :: mode :- ProfileAPI
+  }
+  deriving stock Generic
 
 -- | Endpoints related to account.
 accountEndpoints :: AccountEndpoints Union
-accountEndpoints = AccountEndpoints
-  { _requestCode = requestCodeHandler
-  , _signIn      = signInHandler
-  }
-
-
--- | Handler for sign in first step: when confirmation code is requested.
-requestCodeHandler
-  :: (WithLog m, WithError m, WithSender m) => RequestCodeReq -> m NoContent
-requestCodeHandler RequestCodeReq {..} = do
-  unless (isPhoneValid rc_reqPhone) . throwError Info $ BadRequest
-    "Provided phone is not valid"
-  logInfo $ show rc_reqPhone <> " requested OTP to sign in"
-  account <- whenNothingM (findAccount rc_reqPhone) $ do
-    logDebug $ "Account " <> show rc_reqPhone <> " was not found, creating"
-    createAccount rc_reqPhone
-  void $ sendConfirmationCode SignIn account
-  pure NoContent
-
--- | Handler for sign in first step: when confirmation code is provided.
-signInHandler
-  :: (WithLog m, WithError m, WithJwt m, WithSender m)
-  => SockAddr
-  -- ^ Request address to record 'AuthLog'
-  -> Maybe UserAgent
-  -> SignInReq
-  -> m SignInResp
-signInHandler address agent SignInReq {..} = do
-  logDebug
-    $  show si_reqPhone
-    <> " attempts to sign in with OTP "
-    <> show si_reqCode
-  account <-
-    throwOnNothingM
-        Info
-        (NotAllowed "Provided phone number or OTP is not valid")
-      $ findAccount si_reqPhone
-  logInfo $ "Account " <> show si_reqPhone <> " signed in"
-  -- if code is not valid `signIn` will throw same error as we do it in case
-  -- when there is no such number
-  SignInResp <$> signIn account si_reqCode address agent
+accountEndpoints =
+  AccountEndpoints { _signIn = signInEndpoints, _profile = profileEndpoints }
